ADDED NISDKExtendedEffects CLASS LIBRARY
Extending the real-time-filter-demo project with a new class library,
NISDKExtendedEffects, which has custom effects.  Currently it just has
effects that mimic functionality in the SDK.  The only purpose is to
show a way to achieve the same effect with custom filter code.  You
wouldn't actually want to use the effects that only mimic the SDK as is,
because they perform much worse.